requirements.txt
stdiomask~=0.0.6
mysql-connector-python~=8.0.22